+++ Assets/Prefabs/NPC.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f01e6c5fc341e5e4b8105d2b1020c14f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
