--- Assets/Scripts/Missions/Mission1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca95ae69387816a4ab02253d0127fdc8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
